Refactored code for modularity purposes - Created a `base_scraper.py` which holds the reusable classes if other   index of schools wants to be scraped - Moved the logic of the texas scraper and enricher, implementation to the `texas.py` - Added a `main.py` file which shows how to run the code for each   implementation.
