+++ src/texas_schools_basic_data.csv
@@ -23,7 +23,3 @@ Acton Elem,https://txschools.gov/?view=school&id=111901102&lng=en,Granbury ISD,"
 Ada Mae Faubion Elem,https://txschools.gov/?view=school&id=246913102&lng=en,Leander ISD,"1209 Cypress Creek Rd, Cedar Park, TX 78613",Early Education - Grade 5,,,2
 Adam Elem,https://txschools.gov/?view=school&id=101907107&lng=en,Cypress-Fairbanks ISD,"11303 Honeygrove Ln, Houston, TX 77065",Early Education - Grade 5,,,2
 Adams Elem,https://txschools.gov/?view=school&id=126903101&lng=en,Cleburne ISD,"1492 Is Grove Rd, Cleburne, TX 76031",Prekindergarten - Grade 5,,,2
-Adams Elem,https://txschools.gov/?view=school&id=220901161&lng=en,Arlington ISD,"2220 Sherry St, Arlington, TX 76010",Early Education - Grade 6,,,2
-Adams Hill Elem,https://txschools.gov/?view=school&id=015915123&lng=en,Northside ISD,"9627 Adams Hill Dr, San Antonio, TX 78245",Prekindergarten - Grade 5,,,2
-Adelfa Botello Callejo Elem,https://txschools.gov/?view=school&id=057905247&lng=en,Dallas ISD,"7817 Military Pkwy, Dallas, TX 75227",Prekindergarten - Grade 5,,,2
-Adelle Turner Elem,https://txschools.gov/?view=school&id=057905219&lng=en,Dallas ISD,"5505 S Polk St, Dallas, TX 75232",Prekindergarten - Grade 5,,,2
